Projet_8_API/requirements.txt
fastapi==0.128.0
#tensorflow
pandas
fastapi
uvicorn
azure-monitor-opentelemetry
#mlflow